# app/static/swagger.yaml
swagger: "2.0"
info:
 title: "Auto Shop API"
 description: "API for managing an auto repair shop's mechanics, service tickets, and inventory"
 version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
 - "http"
consumes:
 - "application/json"
produces:
 - "application/json"
...
securityDefinitions:
 bearerAuth:
  type: apiKey
  name: Authorization
  in: header
  description: "JWT Authentication. Example: 'Bearer {token}'"
definitions:
 LoginCredentials:
  type: "object"
  properties:
   email:
    type: "string"
    format: "email"
    example: "admin@test.com"
   password:
    type: "string"
    format: "password"
    example: "admin123"
  required:
   - email
   - password
...
 RegisterUser:
  type: "object"
  properties:
   name:
    type: "string"
    example: "John Doe"
   email:
    type: "string"
    format: "email"
    example: "john@example.com"
   password:
    type: "string"
    format: "password"
    example: "securepass123"
   phone:
    type: "string"
    example: "123-456-7890"
  required:
   - name
   - email
   - password
 UserResponse:
  type: "object"
  properties:
   id:
    type: "integer"
    example: 1
   name:
    type: "string"
    example: "John Doe"
   email:
    type: "string"
    format: "email"
    example: "john@example.com"
   phone:
    type: "string"
    example: "123-456-7890"
   service_tickets:
    type: "array"
    items:
     type: "object"
     properties:
      id:
       type: "integer"
       example: 1
      VIN:
       type: "string"
       example: "1HGCM82633A123456"
      status:
       type: "string"
       enum: ["pending", "in_progress", "completed"]
 ServiceTicketBase:
  type: "object"
  properties:
   VIN:
    type: "string"
    example: "1HGCM82633A123456"
   description:
    type: "string"
    example: "Oil change and tire rotation"
   service_date:
    type: "string"
    format: "date"
    example: "2024-02-15"
   mechanic_ids:
    type: "array"
    items:
     type: "integer"
    example: [1, 2]
   part_ids:
    type: "array"
    items:
     type: "integer"
    example: [1, 3]
  required:
   - VIN
   - description
   - service_date
 ServiceTicketResponse:
  type: "object"
  properties:
   id:
    type: "integer"
    example: 1
   VIN:
    type: "string"
    example: "1HGCM82633A123456"
   description:
    type: "string"
    example: "Oil change and tire rotation"
   service_date:
    type: "string"
    format: "date"
    example: "2024-02-15"
   user:
    type: "object"
    properties:
     id:
      type: "integer"
      example: 1
     name:
      type: "string"
      example: "John Doe"
     email:
      type: "string"
      format: "email"
      example: "john@example.com"
   mechanics:
    type: "array"
    items:
     type: "object"
     properties:
      id:
       type: "integer"
       example: 1
      name:
       type: "string"
       example: "John Smith"
   parts:
    type: "array"
    items:
     type: "object"
     properties:
      id:
       type: "integer"
       example: 1
      name:
       type: "string"
       example: "Oil Filter"
      price:
       type: "number"
       format: "float"
       example: 15.99
...
paths:
 /users/login:
  post:
   tags:
    - "Users"
   summary: "Login endpoint (generates token)"
   parameters:
    - in: "body"
      name: "credentials"
      required: true
      schema:
       $ref: "#/definitions/LoginCredentials"
   responses:
    200:
     description: "Login successful"
     schema:
      type: "object"
      properties:
       token:
        type: "string"
        example: "eyJ0eXAiOiJKV1QiLCJhbGc..."
       message:
        type: "string"
        example: "Successfully logged in"
    401:
     description: "Invalid credentials"
     schema:
      type: "object"
      properties:
       message:
        type: "string"
        example: "Invalid credentials"
    500:
     description: "Internal server error"
     schema:
      type: "object"
      properties:
       message:
        type: "string"
        example: "Internal server error"
 /users:
  post:
   tags:
    - "Users"
   summary: "Register new user"
   description: "Create a new user account. Email must be unique."
   parameters:
    - in: "body"
      name: "user"
      description: "User registration information"
      required: true
      schema:
       $ref: "#/definitions/RegisterUser"
   responses:
    201:
     description: "User created successfully"
     schema:
      type: "object"
      properties:
       message:
        type: "string"
        example: "Successfully registered"
       token:
        type: "string"
        example: "eyJ0eXAiOiJKV1QiLCJhbGc..."
       user:
        $ref: "#/definitions/UserResponse"
    409:
     description: "Email already exists"
     schema:
      type: "object"
      properties:
       message:
        type: "string"
        example: "Email already registered"

  get:
   tags:
    - "Users"
   summary: "Get all users"
   description: "Retrieve list of all users. Requires admin privileges."
   security:
    - bearerAuth: []
   responses:
    200:
     description: "Successfully retrieved users"
     schema:
      type: "array"
      items:
       $ref: "#/definitions/UserResponse"
    401:
     description: "Unauthorized - Invalid or missing token"
    403:
     description: "Forbidden - Not an admin user"
 /users/{id}:
  parameters:
   - in: path
     name: id
     type: integer
     required: true
     description: "User ID"
  get:
   tags:
    - "Users"
   summary: "Get user by ID"
   description: "Retrieve a specific user's information"
   security:
    - bearerAuth: []
   responses:
    200:
     description: "Successfully retrieved user"
     schema:
      $ref: "#/definitions/UserResponse"
    404:
     description: "User not found"
  put:
   tags:
    - "Users"
   summary: "Update user"
   description: "Update user information. Users can only update their own information unless admin."
   security:
    - bearerAuth: []
   parameters:
    - in: "body"
      name: "user"
      description: "Updated user information"
      required: true
      schema:
       $ref: "#/definitions/RegisterUser"
   responses:
    200:
     description: "User updated successfully"
     schema:
      $ref: "#/definitions/UserResponse"
    403:
     description: "Forbidden - Cannot update other users"
    404:
     description: "User not found"
  delete:
   tags:
    - "Users"
   summary: "Delete user"
   description: "Delete a user account. Users can only delete their own account unless admin."
   security:
    - bearerAuth: []
   responses:
    204:
     description: "User deleted successfully"
    403:
     description: "Forbidden - Cannot delete other users"
    404:
     description: "User not found"
 /service-tickets:
  post:
   tags:
    - "Service Tickets"
   summary: "Create new service ticket"
   description: "Create a new service ticket with optional mechanic and part assignments"
   security:
    - bearerAuth: []
   parameters:
    - in: "body"
      name: "ticket"
      description: "Service ticket information"
      required: true
      schema:
       $ref: "#/definitions/ServiceTicketBase"
   responses:
    201:
     description: "Service ticket created successfully"
     schema:
      $ref: "#/definitions/ServiceTicketResponse"
    400:
     description: "Invalid input"
     schema:
      type: "object"
      properties:
       message:
        type: "string"
        example: "One or more mechanic IDs are invalid"
    401:
     description: "Unauthorized - Invalid or missing token"

  get:
   tags:
    - "Service Tickets"
   summary: "Get all service tickets"
   description: "Get paginated list of service tickets for the authenticated user"
   security:
    - bearerAuth: []
   parameters:
    - in: "query"
      name: "page"
      type: "integer"
      description: "Page number"
      default: 1
    - in: "query"
      name: "per_page"
      type: "integer"
      description: "Items per page"
      default: 10
   responses:
    200:
     description: "Successfully retrieved tickets"
     schema:
      type: "object"
      properties:
       tickets:
        type: "array"
        items:
         $ref: "#/definitions/ServiceTicketResponse"
       total:
        type: "integer"
        example: 50
       pages:
        type: "integer"
        example: 5
       current_page:
        type: "integer"
        example: 1
       per_page:
        type: "integer"
        example: 10
       has_next:
        type: "boolean"
        example: true
       has_prev:
        type: "boolean"
        example: false
...
 /mechanics/tickets:
  get:
   tags:
    - "Mechanics"
   summary: "Get mechanic's assigned tickets"
   description: "Get all service tickets assigned to the authenticated mechanic"
   security:
    - bearerAuth: []
   responses:
    200:
     description: "Successfully retrieved tickets"
     schema:
      type: "object"
      properties:
       tickets:
        type: "array"
        items:
         type: "object"
         properties:
          id:
           type: "integer"
           example: 1
          VIN:
           type: "string"
           example: "1HGCM82633A123456"
          description:
           type: "string"
           example: "Oil change and tire rotation"
          service_date:
           type: "string"
           format: "date"
           example: "2024-02-15"
    401:
     description: "Unauthorized - Invalid or missing token"
    404:
     description: "Mechanic not found"
...